+++ Snowflake.Data.Tests/AuthenticationTests/OktaAuthorizationCodeTest.cs
@@ -94,6 +94,8 @@ namespace Snowflake.Data.AuthenticationTests
             parameters.Add(SFSessionProperty.POOLINGENABLED, "false");
             parameters[SFSessionProperty.CLIENT_STORE_TEMPORARY_CREDENTIAL] = "true";
             var host = parameters[SFSessionProperty.HOST];
+            var idpUrl = parameters.TryGetValue(SFSessionProperty.OAUTHTOKENREQUESTURL, out var tokenUrl) ? tokenUrl : string.Empty;
+            var role = parameters.TryGetValue(SFSessionProperty.ROLE, out var roleVal) ? roleVal : string.Empty;
 
             _connectionString = AuthConnectionString.ConvertToConnectionString(parameters);
             Thread connectThread = authTestHelper.GetConnectAndExecuteSimpleQueryThread(_connectionString);
@@ -111,8 +113,8 @@ namespace Snowflake.Data.AuthenticationTests
             }
             finally
             {
-                authTestHelper.RemoveTokenFromCache(host, _login, TokenType.OAuthAccessToken);
-                authTestHelper.RemoveTokenFromCache(host, _login, TokenType.OAuthRefreshToken);
+                authTestHelper.RemoveTokenFromCache(idpUrl, host, _login, role, TokenType.OAuthAccessToken);
+                authTestHelper.RemoveTokenFromCache(idpUrl, host, _login, role, TokenType.OAuthRefreshToken);
             }
         }
     }